1、文章分类 赋值 入数据库
2、文章删除 刷新列表 删除文件
3、文章编辑 分类赋值
4、数据库基层操作类修改

+++ WebLogic/Service/System/FileInfoLogic.cs
@@ -26,16 +26,23 @@ namespace WebLogic.Service.System
         {
             Dictionary<string, object> f = this.dao.GetOne(fileId);
 
-            string filePath = WebPageCore.GetMapPath(f["filePath"].ToString() + f["fileName"].ToString() + "." + f["extName"].ToString());
-
-            if (File.Exists(filePath))
+            if (f != null && f.Count > 0)
             {
-                File.Delete(filePath);
+                string filePath = WebPageCore.GetMapPath(f["filePath"].ToString() + f["fileName"].ToString() + "." + f["extName"].ToString());
+
+                if (File.Exists(filePath))
+                {
+                    File.Delete(filePath);
+                }
+
+                this.dao.Delete(fileId);
+
+                return File.Exists(filePath);
             }
-
-            this.dao.Delete(fileId);
-
-            return File.Exists(filePath);
+            else
+            {
+                return true;
+            }
         }
 
         public string GetFileIds(string[] filePaths)
