+++ WedDao/Dao/Info/RelationshipDao.cs
@@ -37,26 +37,36 @@ namespace WedDao.Dao.Info
 
         public bool SaveList(Int64[] cateIds, Int64 newsId)
         {
-            this.sql = @"delete from [Info_Relationship] where [newsId]=@newsId;";
-
-            this.param = new Dictionary<string, object>();
-            this.param.Add("newsId", newsId);
-            
-            this.db.Update(this.sql, this.param);
-
-            this.sql = @"insert into [Info_Relationship] ([cateId],[newsId])values(@cateId,@newsId)";
-            List<Dictionary<string, object>> paramList = new List<Dictionary<string, object>>();
-
-            for (int i = 0, j = cateIds.Length; i < j; i++)
+            if (newsId > 0)
             {
+                this.sql = @"delete from [Info_Relationship] where [newsId]=@newsId;";
+
                 this.param = new Dictionary<string, object>();
-                this.param.Add("cateId", cateIds[i]);
                 this.param.Add("newsId", newsId);
 
-                paramList.Add(this.param);
+                this.db.Update(this.sql, this.param);
             }
 
-            return this.db.Batch(this.sql, paramList);
+            if (cateIds.Length > 0)
+            {
+                this.sql = @"insert into [Info_Relationship] ([cateId],[newsId])values(@cateId,@newsId)";
+                List<Dictionary<string, object>> paramList = new List<Dictionary<string, object>>();
+
+                for (int i = 0, j = cateIds.Length; i < j; i++)
+                {
+                    this.param = new Dictionary<string, object>();
+                    this.param.Add("cateId", cateIds[i]);
+                    this.param.Add("newsId", newsId);
+
+                    paramList.Add(this.param);
+                }
+
+                return this.db.Batch(this.sql, paramList);
+            }
+            else
+            {
+                return false;
+            }
         }
     }
 }
